SEN-603 Change the "next video" order
When a non-featured video plays, the next video will be based on upload date, regardless of category. This makes sense on mobile because of the feed design, but not on web. Since we’re implementing carousels for the categories too, the next video will be the last uploaded of that category. If none exist, show a replay icon instead of “CONTINUE” (in Intermission Screen).